pyproject.toml
[project]
name = "cse7319-term-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.117.1",
    "i>=2.1.0",
    "ipykernel>=6.30.1",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.12",
    "pandas>=2.3.2",
    "praw>=7.8.1",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.5",
    "uvicorn>=0.37.0",
]

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
]
